test: Move to isolate files with related exercises on components
